+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # GenericsAPI release notes
 =========================================
+1.0.29
+add link associate matrix object feature for PCA app
+add warning message in case of duplicate column header for matrix importer
 
 1.0.28
 add ubiquity filtering operation to the transform matrix app